Make react-hot-loader optimise away when DefinePlugin sets NODE_ENV to production, and do not required NODE_ENV to be set in webpack itself
