+++ babel.js
@@ -1,7 +1,3 @@
 'use strict'
 
-if (process.env.NODE_ENV === 'production') {
-  module.exports = require('./dist/babel.production.min.js');
-} else {
-  module.exports = require('./dist/babel.development.js');
-}
+module.exports = require('./dist/babel.js');
