+++ src/babel.js
@@ -1,5 +1,7 @@
 import { REGENERATE_METHOD } from './internal/constants'
 
+const RHLPackage = 'react-hot-loader'
+
 const templateOptions = {
   placeholderPattern: /^([A-Z0-9]+)([A-Z0-9_]+)$/,
 }
@@ -27,37 +29,61 @@ module.exports = function plugin(args) {
   }
   const { types: t, template } = args
 
+  // Prevent our own rollup compilation substituting this
+  const processEnvNodeEnv = 'process.' + 'env.' + 'NODE_ENV'; // eslint-disable-line no-useless-concat
+
   const buildRegistration = template(
     'reactHotLoader.register(ID, NAME, FILENAME);',
     templateOptions,
   )
   const headerTemplate = template(
-    `(function () {
-       var enterModule = require('react-hot-loader').enterModule;
-       enterModule && enterModule(module);
-     }())`,
+    `
+(function () {
+  // Allow tree shaking to drop the react-hot-loader module by wrapping with a conditional that DefinePlugin can replace
+  // This cannot be a ternary - must be an IF
+  if (${processEnvNodeEnv} != 'production') {
+    var enterModule = require('react-hot-loader').enterModule;
+    enterModule && enterModule(module);
+  }
+}())
+    `,
     templateOptions,
   )
   const evalTemplate = template('this[key]=eval(code);', templateOptions)
 
-  // We're making the IIFE we insert at the end of the file an unused variable
-  // because it otherwise breaks the output of the babel-node REPL (#359).
   const buildTagger = template(
     `
 (function () {
-  var reactHotLoader = require('react-hot-loader').default;
-  var leaveModule = require('react-hot-loader').leaveModule;
+  // Allow tree shaking to drop the react-hot-loader module by wrapping with a conditional that DefinePlugin can replace
+  // This cannot be a ternary - must be an IF
+  if (${processEnvNodeEnv} != 'production') {
+    var reactHotLoader = require('react-hot-loader').default;
+    var leaveModule = require('react-hot-loader').leaveModule;
 
-  if (!reactHotLoader) {
-    return;
+    if (!reactHotLoader) {
+      return;
+    }
+
+    REGISTRATIONS
+
+    leaveModule(module);
   }
-
-  REGISTRATIONS
-
-  leaveModule(module);
 }());
-  `,
+    `,
     templateOptions,
+  )
+
+  const hotTemplate = template(
+    `
+    function(){
+      if (${processEnvNodeEnv} == 'production') {
+        return ARGUMENTNODE
+      } else {
+        return HOTNODE
+      }
+    }()
+    `,
+    templateOptions
   )
 
   // Gather top-level variables, functions, and classes.
@@ -82,9 +108,116 @@ module.exports = function plugin(args) {
     }
   }
 
+  function isImportedFromRHL(path, name) {
+    const binding = path.scope.getBinding(name)
+    const bindingType = binding && binding.path.node.type
+
+    if (
+      bindingType === 'ImportSpecifier' ||
+      bindingType === 'ImportNamespaceSpecifier'
+    ) {
+      const bindingParent = binding.path.parent
+      return bindingParent.source.value === RHLPackage
+    }
+
+    return false
+  }
+
+  function getRHLContext(file) {
+    const context = []
+    const { body } = file.ast.program
+
+    for (let i = 0; i < body.length; i++) {
+      const bodyItem = body[i]
+      const { source, specifiers } = bodyItem
+
+      if (bodyItem.type !== 'ImportDeclaration' || source.value !== RHLPackage) {
+        continue
+      }
+
+      for (let j = 0; j < specifiers.length; j++) {
+        const specifier = specifiers[j]
+
+        if (specifier.type === 'ImportNamespaceSpecifier') {
+          context.push({
+            kind: 'namespace',
+            local: specifier.local.name,
+          })
+        } else if (specifier.type === 'ImportSpecifier') {
+          const specifierData = {
+            kind: 'named',
+            local: specifier.local.name,
+          }
+
+          if (specifier.imported.name === 'hot') {
+            context.push(specifierData)
+          }
+        }
+      }
+    }
+
+    return context.length ? context : null
+  }
+
+  function replaceHotPath(hotPath, argumentNode, hotName) {
+    // Prevent recursion - when we replaced we re-added hot() again so do not replace it again
+    if (hotPath.node.reactHotLoaderReplaced) {
+      return
+    }
+    hotPath.node.reactHotLoaderReplaced = true
+    // Replace the hot call with a conditional expression that UglifyJs 3 will completely remove
+    // UglifyJs 2 will collapse it to a simple return
+    hotPath.replaceWith(hotTemplate({ ARGUMENTNODE: argumentNode.arguments[0], HOTNODE: hotPath.node}))
+  }
+
   const REGISTRATIONS = Symbol('registrations')
   return {
+    pre() {
+      // ignore hot replacement in files that do not use RHL
+      this.rhlContext = getRHLContext(this.file)
+      this.cancelHotReplacement = !this.rhlContext
+    },
     visitor: {
+      CallExpression(path) {
+        if (this.cancelHotReplacement) {
+          return
+        }
+
+        for (let i = 0; i < this.rhlContext.length; i++) {
+          const specifier = this.rhlContext[i]
+
+          if (specifier.kind === 'named') {
+            if (
+              path.node.callee.name === specifier.local &&
+              // ensure that this is `hot` from RHL
+              isImportedFromRHL(path, specifier.local) &&
+              path.parent.type === 'CallExpression' &&
+              path.parent.arguments[0] &&
+              path.parent.arguments[0].type === 'Identifier'
+            ) {
+              replaceHotPath(path.parentPath, path.parent, specifier.local);
+              break
+            }
+          } else if (specifier.kind === 'namespace') {
+            if (
+              path.node.callee.callee &&
+              path.node.callee.callee.type === 'MemberExpression' &&
+              path.node.callee.callee.object.type === 'Identifier' &&
+              path.node.callee.callee.object.name === specifier.local &&
+              // ensure that this is from RHL
+              isImportedFromRHL(path, specifier.local) &&
+              path.node.callee.callee.property.type === 'Identifier' &&
+              path.node.callee.callee.property.name === 'hot' &&
+              path.node.arguments[0] &&
+              path.node.arguments[0].type === 'Identifier'
+            ) {
+              replaceHotPath(path, path.node, specifier.local)
+              break
+            }
+          }
+        }
+      },
+
       ExportDefaultDeclaration(path, state) {
         const { file } = state
         // Default exports with names are going
